api/requirements/base.txt
Flask
Flask-RESTful
python-dotenv
pymongo==3.9.0
flask-mongoengine==0.9.5
mongoengine==0.18.2
marshmallow==2.20.0
Flask-Cors
python-slugify
